feat(cqrs): moved cqrs traits to crate
